The new home merges into FieldDesk's desktop branch, under the roster
Owner ruling 2026-08-02. `RootLanding` is `user ? <FieldDesk /> : <Home />`, so
a signed-in viewer never reaches `Home` — everything #1557 built there was dead
code. Flipping the route instead would have orphaned the three things only
FieldDesk's desktop branch offers: switching the carried character, "begin a new
self", and the Albescent invitation.

So `Home.tsx` goes back to guest-only, byte-for-byte as it was on main, and
`ContinueSection` / `ContinueRow` / `BrowseSection` move into FieldDesk's DESKTOP
branch (the mobile skins dispatch away above it, and stay #1554's).

They sit BELOW the roster. The roster answers "who am I playing" and the new
sections answer "what do I do now", and a returning player has to settle the
first before the second is addressed to anyone: both "your drafts" and "tasks
you can sign up for" are scoped to the carried life, so a viewer who is about to
switch lives is being shown the wrong ones.

The copy stays in `home.json` under `signedIn.*` — this component is what `/`
renders for a signed-in viewer, and the locales README gives home.json "Home /
landing page copy". `common:fieldDesk.*` stays the roster's and the mobile
skins'.

The 8 tests move with the code to
`pages/fieldDesk/__tests__/fieldDeskDesktopHome.test.tsx`. The one that asserted
which of `/`'s two screens rendered is now redundant with
`__tests__/rootLandingChunks.test.tsx`; in its place is the assertion this
rework exists for — the roster is still there, and still above.
